backend: set pixels per point calculated from monitor
Use the glfw monitor to find the pixels per point. The assumption is
pixels per point is 1.0 is when the pixels per inch is 96.0

This may not be the correct method but hope it is.

This also doesn't work with multiple monitors since only the primary
monitor is used to get the information which is not desirable. If it
is possible to get the current monitor on which the window resides,
then this becomes a trivial task.