backend: shader file path based on crate file path
When using relative path, the environment path is used. So when this
crate is used as a dependency in another crate, it causes the paths to
be invalid.

To fix this, find the crate root file path by using `file!()` to get
the current file's file path and then take it's parent the correct
number of times. To this, the relative path can be joined.