allow multiple versions of egui through features with 0.16 changes
This allows the user to choose the version of egui to use through
features, default is the latest which is 0.16 as of right now.

0.16 also introduced many breaking changes, so update where needed
through cfg(feature) to allow for 0.14, 0.15 and 0.16 of egui. This
may not be the best idea since many more breaking changes later, the
code to support all these changes will make it basically impossible to
maintain. Might make sense to scrap the multiple versions of egui
through features :( It is a good test though.