cargo: release 0.6.0
BREAKING:
* Upgraded egui to 0.23
* Pixels per point is supported now. This is considered breaking
because the previous implementation made it such that the default
pixels per point even if set would result in scaling of 1.0. This is
generally the default for most systems, so it should not affect most
systems but for few with custom scaling options, it may cause the UI
to look larger or smaller.

Non breaking:
* Internal improvements
  + No longer uses projection matrix, thus reducing the cost of the
  vertex shader.
  + Shader variables have been renamed for better clarity