cargo: update glfw to 0.58
For some reason using `0.*` doesn't mean that the dependent decides
the `glfw` version. It always only picks the latest even if the
dependent specifies `glfw` version to be less than the latest. It
might have to do with semver compatibility coming into consideration
when resolving `glfw`.

NOTE: when testing, forcing `egui_glfw` to use `0.57` but the
dependent to use `0.58`, `egui_glfw`'s `glfw` was not resolved to
`0.58` but remained at `0.57`. This obviously leads to compilation
errors because the `glfw` types used in `egui_glfw`'s public API has a
different version compared to the `glfw` types in the dependent. Does
semver consider `0.x` to `0.y` as a breaking change, so it doesn't
resolve it?