texture: TextureRGBA8: texture_options
Use the [`egui::TextureOptions`] in
[`egui::epaint::image::ImageDelta`] for the sampling parameters.

TODO: this makes the text blurry again because the default egui
texture (the one containing the font book) is sampled with
`gl::LINEAR`. This likely means that there is some kind of half pixel
offset of similar that needs to be looked into. In
`examples/demo_app`'s [`egui_demo_lib::ColorTest`], the 1 pixel wide
lines are not pure white and pure black but rather a hint of grey.