texture: remove TODO since blurry text caused on too many MSAA samples
The cause for the blurry text is due to the MSAA samples being too
high. [`egui`] does not require MSAA since egui already does some form
of antialiasing during tessellation. See
[`egui::epaint::tessellator::TessellationOptions::featuring`] for
details about antialiasing that egui performs.

Even though egui itself doesn't require MSAA with the default
settings, it is not possible to disable MSAA, since feathering might
be disabled so it would be up to MSAA to do the antialiasing.

Now for the reason why too many MSAA samples causes blurry
text. OpenGL on my system (NVIDIA GTX 1650, driver version 525.147.05)
reports the max supported samples as 32 (determined using
`gl::GetIntegerv(gl::MAX_SAMPLES, &mut max_samples)`). Even though the
requested MSAA samples is 16, lower than 32, it causes blurry
text. The vulkan specification has support for only up to 16
samples (32 and 64 are supported but it is not standard). So after
checking what vulkan reports, max supported samples is 8, it is
obvious that OpenGL is reporting something that isn't supported by the
hardware. But why is the question.

There is NV_internalformat_sample_query
<https://registry.khronos.org/OpenGL/extensions/NV/NV_internalformat_sample_query.txt>
extension by NVIDIA that allows for reporting of a larger number of
samples than supported by hardware by doing supersampling for the
larger number of samples.

```
    Some NVIDIA drivers support multisample modes which are internally
    implemented as a combination of multisampling and automatic
    supersampling in order to obtain a higher level of anti-aliasing than
    can be directly supported by hardware.
```

This means that at 16 or 32 samples (8 is definitely support on my
hardware because vulkan reports it as supported and because it can be
verified based on if it causes blurriness), it is actually
supersampling that is taking place (along with MSAA). So when
downsampling the supersampled texel, it seems to be using some sort of
linear filtering method, which means even if the mesh has an exact
boundary (perfectly vertical or horizontal lines), it leads to a blend
at this edge boundary, leading to blurry text and lines (and shapes).

So what can be done about this now? There isn't any good way to tackle
this problem since OpenGL cannot request for true number of MSAA
samples. So the only option is to ensure that the user is aware of
this limitation and does not use too many MSAA samples. Since OpenGL
guarantees 4 samples must be supported, it is a good default to
use (if MSAA is needed in the first place, UI only tools shouldn't,
tools using OpenGL for rendering other things might want to use MSAA).